Criado classe de erros personalizada

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,17 @@
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo com esse nome.");
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
